chore(structure): reorganize project by separating frontend and backend folders
