+++ .gitignore
@@ -1,0 +1,26 @@
+node_modules
+# Node modules in root (if any)
+node_modules/
+
+# Logs
+npm-debug.log*
+yarn-debug.log*
+lerna-debug.log*
+
+# OS files
+.DS_Store
+Thumbs.db
+
+# Editor folders and files
+.vscode/
+.idea/
+*.sublime-project
+*.sublime-workspace
+
+# Environment files (if any)
+.env
+.env.local
+.env.*.local
+
+# Misc
+*.log
