+++ frontend/.gitignore
@@ -1,0 +1,46 @@
+# Dependencies
+node_modules/
+.pnp/
+.pnp.js
+
+# Build output
+build/
+dist/
+
+# Vite cache
+.vite/
+
+# Testing
+coverage/
+
+# Production
+.env.production
+.env.local
+.env.development.local
+.env.test.local
+
+# Logs
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+
+# OS
+.DS_Store
+Thumbs.db
+
+# IDEs
+.vscode/
+.idea/
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+
+# Misc
+*.log
+*.tgz
+.cache/
+*.zip
+*.env
+.env.*
