+++ frontend/src/components/SettingsPanel.tsx
