+++ frontend/src/components/StreamCard.tsx
