+++ frontend/src/components/StreamGrid.tsx
