+++ package.json
@@ -2,53 +2,21 @@
   "name": "livestream-dashboard",
   "version": "1.0.0",
   "private": true,
-  "dependencies": {
-    "@types/node": "^16.18.0",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "@types/styled-components": "^5.1.26",
-    "hls.js": "^1.4.0",
-    "i18next": "^22.4.15",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-grid-layout": "^1.3.4",
-    "react-i18next": "^12.2.2",
-    "react-icons": "^4.8.0",
-    "react-scripts": "5.0.1",
-    "socket.io-client": "^4.7.2",
-    "styled-components": "^5.3.9",
-    "typescript": "^4.9.5",
-    "uuid": "^9.0.0",
-    "express": "^5.1.0",
-    "socket.io": "^4.8.1"
-  },
   "scripts": {
-    "start": "concurrently \"react-scripts start\" \"node backend/index.js\"",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "start": "concurrently \"npm run frontend\" \"npm run backend\"",
+    "dev": "npm run start",
+
+    "frontend": "npm --prefix frontend start",
+    "backend": "npm --prefix backend start",
+
+    "build": "npm --prefix frontend build",
+    "test": "npm --prefix frontend test",
+
+    "install:frontend": "npm --prefix frontend install",
+    "install:backend": "npm --prefix backend install",
+    "install:all": "npm install && npm run install:frontend && npm run install:backend"
   },
   "devDependencies": {
-    "@types/react-grid-layout": "^1.3.5",
-    "@types/uuid": "^9.0.2",
     "concurrently": "^9.1.2"
   }
 }
